Big calendars/addressbooks are really slow and requests require a lot of memory. The main reason is vobject. Parsing and serializing with it is slow and it's a memory hog.
This PR reduces the usage of vobject by caching serialization results and metadata for pre-filtering on disk. The memory usage gets further reduced by only keeping one item in memory at a time.

All commits until ``Cache metadata when storage is read-only`` are general improvements. They fix some small bugs, reduce the memory usage by not loading all items at once into memory and fix the slowness that was introduced by collision detection for file system paths.
The following commits improve speed by:
  * Caching the serialization of the whole collection on disk.
  * Caching the serialization of items and metadata (tag, start and end time) for pre-filtering on disk.
  * Caching the etag and metadata of the collection for the duration of the request in memory, if the access is read-only.

This PR is also merged into https://github.com/Unrud/Radicale.

## Benchmark
The tests are performed on a calendar and addressbook with 10,000 files. The approximate memory usage of the requests is noted in brackets.

### Calendar
```
Type:   PROPFIND_getetag (initialize cache)
Before: 15.5 seconds (138,156 kB)
After:  17.8 seconds (    372 kB)

Type:   PROPFIND_getetag
Before: 15.3 seconds (136,604 kB)
After:   0.5 seconds (      0 kB)

Type:   REPORT_calendar-query_getetag
Before: 16.8 seconds (154,180 kB)
After:   1.0 seconds ( 21,964 kB)
PR #586: 1.0 seconds ( 15,356 kB)

Type:   REPORT_calendar-multiget_getetag
Before: 16.4 seconds ( 28,248 kB)
After:   2.1 seconds ( 28,660 kB)
PR #586: 2.2 seconds ( 19,852 kB)

Type:   REPORT_calendar-query_filter-all_getetag
Before: 18.3 seconds (154,448 kB)
After:   1.1 seconds ( 22,332 kB)
PR #586: 1.1 seconds ( 15,092 kB)

Type:   REPORT_calendar-query_filter-none_getetag
Before: 11.0 seconds (133,580 kB)
After:   0.5 seconds (      0 kB)

Type:   GET
Before: 30.5 seconds (138,756 kB)
After:   1.2 seconds (  2,740 kB)
```

### Addressbook
```
Type:   PROPFIND_getetag (initialize cache)
Before:  8.0 seconds ( 69,732 kB)
After:   9.4 seconds (    284 kB)

Type:   PROPFIND_getetag
Before:  8.0 seconds ( 70,020 kB)
After:   0.5 seconds (      0 kB)

Type:   GET
Before: 16.0 seconds ( 70,012 kB)
After:   1.1 seconds (  3,692 kB)
```
**Note:** In all "Before" tests the collision checking for file names was disabled, because it slowed down the requests significantly and made the results less comparable.