@pbiering 2017-03-13
sounds interesting and probably superseed my "simple" in-memory caching, yours is persistent and survives Radicale restarts.