@Unrud 2017-03-18
I've implemented prefiltering for the multifilesystem backend and all collection items are now verified by vobject. For this I added a on-disk cache that stores the etags of verified items and some extracted parameter for prefiltering. Items that are not in the cache are always parsed and serialized before they are returned by **Collection.get**.

Prefiltering takes about 1 second to filter out 10,000 items.

I've added some helper functions for prefiltering to xmlutils. They convert the complicated XML filters to a tag and time range (POSIX timestamps as ints) and can find the tag and an enclosing time range for vobject_items. 

> superseed my "simple" in-memory caching

I did some experiments. I loaded a collection (with simple VEVENTs) that required ~3.4 MiB of space on disk. Caching it in memory required an additional ~128 MiB of RAM.