@Unrud 2017-04-04
I changed the item cache to only keep one cache entry per item. Entries for deleted items get cleaned up automatically.