@Unrud 2017-04-20
Rebased, fixed all flake8 warnings, simplified and reordered the commits and added some comments.

All commits until ``Cache metadata when storage is read-only`` are general improvements. They fix some small, reduce the RAM usage by not loading all items at once into memory and fix the slowness that was introduced by collision detection for file system paths.
The following commits improve speed by:
  * Caching the serialization of the whole collection on disk.
  * Caching the serialization of items and metadata (tag, start and end time) for pre-filtering on disk.
  * Caching the etag and metadata of the collection for the duration of the request in memory, if the access is read-only.