@Unrud 2017-04-20
I ran the benchmark again, to incorporate the new changes. This time I also included the approximate RAM usage of the requests. The tests are performed on a calendar and addressbook with 10,000 files.

## Benchmark
### Calendar
```
Type:   PROPFIND_getetag (initialize cache)
Before: 15.5 seconds (138,156 kB)
After:  17.8 seconds (    372 kB)

Type:   PROPFIND_getetag
Before: 15.3 seconds (136,604 kB)
After:   0.5 seconds (      0 kB)

Type:   REPORT_calendar-query_getetag
Before: 16.8 seconds (154,180 kB)
After:   1.0 seconds ( 21,964 kB)

Type:   REPORT_calendar-multiget_getetag
Before: 16.4 seconds ( 28,248 kB)
After:   2.1 seconds ( 28,660 kB)

Type:   REPORT_calendar-query_filter-all_getetag
Before: 18.3 seconds (154,448 kB)
After:  10.2 seconds ( 22,332 kB)

Type:   REPORT_calendar-query_filter-none_getetag
Before: 11.0 seconds (133,580 kB)
After:   0.5 seconds (      0 kB)

Type:   GET (initialize cache)
Before: 30.4 seconds (138,752 kB)
After:  16.3 seconds (102,228 kB)

Type:   GET
Before: 30.5 seconds (138,756 kB)
After:   0.6 seconds (  6,860 kB)
```

### Addressbook
```
Type:   PROPFIND_getetag (initialize cache)
Before:  8.0 seconds ( 69,732 kB)
After:   9.4 seconds (    284 kB)

Type:   PROPFIND_getetag
Before:  8.0 seconds ( 70,020 kB)
After:   0.5 seconds (      0 kB)

Type:   GET (initialize cache)
Before: 16.0 seconds ( 70,012 kB)
After:   1.1 seconds (  3,836 kB)

Type:   GET
Before: 15.8 seconds ( 69,460 kB)
After:   0.6 seconds (  5,012 kB)
```
**Note:** In all "Before" tests the collision checking for file names was disabled, because it slowed down the requests significantly and made the results less comparable.