@Unrud 2017-05-07
I implemented the serialization of calendars without vobject. The calendar components are merged manually. This improves the time of GET requests from ~30 seconds (~100 MiB of memory) to ~1 second (~5 MiB of memory). Consequentially I removed the cache for the serialization of collections, because it's not needed anymore.

Filtering of items with vobject in ``xmlutils.py`` can now be skipped if the storage backend indicates that it already matched all ``filters`` in prefiltering. This is used to optimize requests with simple filters, that only match the tag and a big time range (like filtering all events from the last three months).
At least DAVdroid and InfCloud only make use of simple filters. This improves the time for the requests from ~10 seconds to ~1 second.

vobject is now only used for one-time input validation and for complex filtering.
I've updated the benchmark in the first post and also included results from #586.

I've created a fork of Radicale at https://github.com/Unrud/Radicale where I merged all my open PRs.