@jvoisin 2017-05-16
Nice!
What is needed to get this merged @liZe ?