@Unrud 2017-05-26
@htgoebel Thanks for reviewing, I will fix the identified problems soon.